pyproject.toml
[tool.poetry]
name = "pythonweb-hw08"
version = "0.1.0"
description = ""
authors = ["Dmytro-Tarasenko <ug.dotenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mongoengine = "^0.28.2"
pydantic = "^2.6.4"
redis = "^5.0.3"
rich = "^13.7.1"
pymongo = "^4.6.2"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
faker = "^24.3.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
